replace broken/obsolete delete-workspaces tool
